Move injection logic
